Unify transform code.
