+++ src/transform.cpp
@@ -18,6 +18,130 @@
 //  along with UWOT.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <Rcpp.h>
+#include <RcppParallel.h>
+// [[Rcpp::depends(RcppParallel)]]
+
+
+
+
+struct AverageWorker : public RcppParallel::Worker {
+
+  const RcppParallel::RMatrix<double> train_embedding;
+  const RcppParallel::RMatrix<int> nn_index;
+  RcppParallel::RMatrix<double> embedding;
+  const size_t nc;
+  const size_t nnbrs;
+  const double one_over_n;
+
+  AverageWorker(Rcpp::NumericMatrix train_embedding, Rcpp::IntegerMatrix nn_index,
+                Rcpp::NumericMatrix embedding
+  ) :
+    train_embedding(train_embedding), nn_index(nn_index),
+    embedding(embedding),
+    nc(train_embedding.ncol()), nnbrs(nn_index.ncol()),
+    one_over_n(1.0 / nnbrs)
+  {  }
+
+  void operator()(std::size_t begin, std::size_t end) {
+    double sumc[nc];
+    for (std::size_t i = begin; i < end; i++) {
+      for (size_t k = 0; k < nc; k++) {
+        sumc[k] = 0.0;
+      }
+
+      for (size_t j = 0; j < nnbrs; j++) {
+        auto nbr = nn_index(i, j) - 1;
+        for (size_t k = 0; k < nc; k++) {
+          sumc[k] += train_embedding(nbr, k);
+        }
+      }
+
+      for (size_t k = 0; k < nc; k++) {
+        embedding(i, k) = sumc[k] * one_over_n;
+      }
+    }
+  }
+};
+
+
+// [[Rcpp::export]]
+Rcpp::NumericMatrix init_transform_av_parallel(Rcpp::NumericMatrix train_embedding,
+                                               Rcpp::IntegerMatrix nn_index,
+                                               const size_t grain_size = 1) {
+  Rcpp::NumericMatrix embedding(nn_index.nrow(), train_embedding.ncol());
+
+  AverageWorker worker(train_embedding, nn_index, embedding);
+
+  RcppParallel::parallelFor(0, nn_index.nrow(), worker, grain_size);
+
+  return embedding;
+}
+
+// [[Rcpp::export]]
+Rcpp::NumericMatrix init_transform_av_cpp(Rcpp::NumericMatrix train_embedding,
+                                          Rcpp::IntegerMatrix nn_index) {
+  Rcpp::NumericMatrix embedding(nn_index.nrow(), train_embedding.ncol());
+
+  AverageWorker worker(train_embedding, nn_index, embedding);
+  worker(0, nn_index.nrow());
+
+  return embedding;
+}
+
+struct WeightedAverageWorker : public RcppParallel::Worker {
+
+  const RcppParallel::RMatrix<double> train_embedding;
+  const RcppParallel::RMatrix<int> nn_index;
+  const RcppParallel::RMatrix<double> nn_weights;
+  RcppParallel::RMatrix<double> embedding;
+  const size_t nc;
+  const size_t nnbrs;
+
+  WeightedAverageWorker(Rcpp::NumericMatrix train_embedding, Rcpp::IntegerMatrix nn_index,
+                        const Rcpp::NumericMatrix& nn_weights, Rcpp::NumericMatrix embedding
+  ) :
+    train_embedding(train_embedding), nn_index(nn_index), nn_weights(nn_weights),
+    embedding(embedding),
+    nc(train_embedding.ncol()), nnbrs(nn_index.ncol())
+  {  }
+
+  void operator()(std::size_t begin, std::size_t end) {
+    double sumc[nc];
+    for (std::size_t i = begin; i < end; i++) {
+      for (size_t k = 0; k < nc; k++) {
+        sumc[k] = 0.0;
+      }
+      double sumw = 0.0;
+
+      for (size_t j = 0; j < nnbrs; j++) {
+        auto nbr = nn_index(i, j) - 1;
+        double w = nn_weights(i, j);
+        sumw += w;
+        for (size_t k = 0; k < nc; k++) {
+          sumc[k] += train_embedding(nbr, k) * w;
+        }
+      }
+
+      for (size_t k = 0; k < nc; k++) {
+        embedding(i, k) = sumc[k] / sumw;
+      }
+    }
+  }
+};
+
+// [[Rcpp::export]]
+Rcpp::NumericMatrix init_transform_parallel(Rcpp::NumericMatrix train_embedding,
+                                            Rcpp::IntegerMatrix nn_index,
+                                            Rcpp::NumericMatrix nn_weights,
+                                            const size_t grain_size = 1) {
+  Rcpp::NumericMatrix embedding(nn_index.nrow(), train_embedding.ncol());
+
+  WeightedAverageWorker worker(train_embedding, nn_index, nn_weights, embedding);
+
+  RcppParallel::parallelFor(0, nn_index.nrow(), worker, grain_size);
+
+  return embedding;
+}
 
 // Initialize embedding as a weighted average of nearest neighbors of each point
 // train_embedding: n_train x dim matrix of final embedding coordinates
@@ -28,66 +152,11 @@
 // [[Rcpp::export]]
 Rcpp::NumericMatrix init_transform_cpp(Rcpp::NumericMatrix train_embedding,
                                        Rcpp::IntegerMatrix nn_index,
-                                       Rcpp::NumericMatrix weights) {
-  auto nr = nn_index.nrow();
-  auto nc = train_embedding.ncol();
-  auto nnbrs = nn_index.ncol();
+                                       Rcpp::NumericMatrix nn_weights) {
+  Rcpp::NumericMatrix embedding(nn_index.nrow(), train_embedding.ncol());
 
-  Rcpp::NumericMatrix embedding(nr, nc);
-
-  double sumc[nc];
-
-  for (auto i = 0; i < nr; i++) {
-    for (auto k = 0; k < nc; k++) {
-      sumc[k] = 0.0;
-    }
-    double sumw = 0.0;
-
-    for (auto j = 0; j < nnbrs; j++) {
-      auto nbr = nn_index.at(i, j) - 1;
-      double w = weights.at(i, j);
-      sumw += w;
-      for (auto k = 0; k < nc; k++) {
-        sumc[k] += train_embedding.at(nbr, k) * w;
-      }
-    }
-
-    for (auto k = 0; k < nc; k++) {
-      embedding.at(i, k) = sumc[k] / sumw;
-    }
-  }
-
-  return embedding;
-}
-
-// [[Rcpp::export]]
-Rcpp::NumericMatrix init_transform_av_cpp(Rcpp::NumericMatrix train_embedding,
-                                          Rcpp::IntegerMatrix nn_index) {
-  auto nr = nn_index.nrow();
-  auto nc = train_embedding.ncol();
-  auto nnbrs = nn_index.ncol();
-
-  Rcpp::NumericMatrix embedding(nr, nc);
-
-  double sumc[nc];
-  double one_over_n = 1.0 / nnbrs;
-
-  for (auto i = 0; i < nr; i++) {
-    for (auto k = 0; k < nc; k++) {
-      sumc[k] = 0.0;
-    }
-
-    for (auto j = 0; j < nnbrs; j++) {
-      auto nbr = nn_index.at(i, j) - 1;
-      for (auto k = 0; k < nc; k++) {
-        sumc[k] += train_embedding.at(nbr, k);
-      }
-    }
-
-    for (auto k = 0; k < nc; k++) {
-      embedding.at(i, k) = sumc[k] * one_over_n;
-    }
-  }
+  WeightedAverageWorker worker(train_embedding, nn_index, nn_weights, embedding);
+  worker(0, nn_index.nrow());
 
   return embedding;
 }
